Interface work for deleting/restoring docs
As both admin and user. Issue #363

+++ app/Http/routes.php
@@ -77,6 +77,7 @@ Route::pattern('user', '[0-9]+');
 Route::pattern('date', '[0-9]+');
 Route::pattern('group', '[0-9]+');
 Route::pattern('state', Doc::validStatesPattern());
+Route::pattern('admin', 'admin');
 
 /*
 *   Route - Model bindings
@@ -232,7 +233,7 @@ Route::post('api/docs/{doc}/content', 'DocumentApiController@postContent')->midd
 
 Route::post('api/docs/{doc}/featured-image', 'DocumentsController@uploadImage')->middleware(['doc.access.edit']);
 Route::post('api/docs/{doc}/status', 'DocumentApiController@postStatus')->middleware(['doc.access.edit']);
-Route::delete('api/docs/{doc}', 'DocumentApiController@deleteDoc')->middleware(['doc.access.edit']);
+Route::delete('api/docs/{doc}/{admin?}', 'DocumentApiController@deleteDoc')->middleware(['doc.access.edit']);
 Route::get('api/docs/{doc}/restore', 'DocumentApiController@getRestoreDoc')->middleware(['doc.access.edit']);
 Route::delete('api/docs/{doc}/featured-image', 'DocumentsController@deleteImage')->middleware(['doc.access.edit']);
 Route::delete('api/docs/{doc}/dates/{date}', 'DocumentApiController@deleteDate')->middleware(['doc.access.edit']);
